Edit Cell.intersects method in tmx.py to support cell based collision
Without accounting for the extra pixel that other.y/x adds, the intersect method
causes a bug that overestimates the collision point for tile based sprites. Subtracting 1 from
each value on the left side of the conditional statment evens this out and fixes it. Not sure if this works
the same for all games, so be aware of this change.